Move recipe modification route and tests to actions module
