+++ meal_planner/routers/actions.py
@@ -10,7 +10,10 @@ from starlette.datastructures import FormData
 from meal_planner.form_processing import _parse_recipe_form_data
 from meal_planner.main import internal_client, rt  # Assuming internal_client is needed
 from meal_planner.models import RecipeBase
+from meal_planner.services.call_llm import generate_modified_recipe
+from meal_planner.services.process_recipe import postprocess_recipe
 from meal_planner.ui.common import CSS_ERROR_CLASS, CSS_SUCCESS_CLASS
+from meal_planner.ui.edit_recipe import build_modify_form_response
 
 logger = logging.getLogger(__name__)
 
@@ -84,4 +87,155 @@ async def post_save_recipe(request: Request):
                 headers={"HX-Trigger": "recipeListChanged"},
             )
 
+    return result
+
+
+@rt("/recipes/modify")
+async def post_modify_recipe(request: Request):
+    """
+    Handles recipe modification requests from the recipe editing UI.
+
+    This endpoint processes form data containing a current recipe, an original
+    recipe (for diffing and reference), and an optional user-provided modification
+    prompt. It orchestrates recipe data parsing, calls an LLM service for
+    modifications if requested, and then re-renders the entire recipe edit form UI
+    with any relevant messages.
+
+    Expected form fields:
+    - Current recipe: 'name', 'ingredients' (list), 'instructions' (list).
+    - Original recipe (for diff): 'original_name', 'original_ingredients' (list),
+      'original_instructions' (list).
+    - AI modification: 'modification_prompt' (string, optional).
+
+    Workflow and Error Handling:
+
+    All responses from this endpoint will be HTTP 200 OK and will contain
+    the full recipe modification form, updated as necessary.
+
+    1.  **Initial Form Data Parsing and Validation:**
+        -   The endpoint first parses all submitted recipe data.
+        -   If essential data for the current or original recipe is missing or
+            structurally invalid (e.g., fails Pydantic `RecipeBase` validation),
+            the form is re-rendered with the error message "Invalid recipe data.
+            Please check the fields." The submitted form data is used to
+            repopulate the fields as much as possible.
+
+    2.  **Modification Prompt Handling:**
+        -   If no 'modification_prompt' is provided by the user (after successful
+            initial parsing), the form is re-rendered with the error message
+            "Please enter modification instructions." The current recipe data
+            remains in the form.
+        -   If a 'modification_prompt' is provided:
+            -   The LLM service is called to generate a modified recipe.
+            -   **On Successful LLM Modification:**
+                -   The form is re-rendered, showing the new `modified_recipe`
+                  as current, with no error message.
+            -   **On LLM Service or Postprocessing Error (`RuntimeError`):**
+                -   The error is logged.
+                -   The form is re-rendered with the specific error message from
+                    `RuntimeError`. The recipe data from *before* the
+                    LLM call is used to populate the form.
+
+    3.  **Other Unexpected Internal Errors (Generic `Exception`):**
+        -   If any other unexpected `Exception` occurs:
+            -   The error is logged with full details.
+            -   The form is re-rendered with a generic "Critical Error..."
+                message. The recipe data from before the failed operation is used.
+
+    Args:
+        request: The FastAPI request object, containing the form data.
+
+    Returns:
+        A `Group` of components representing the full recipe modification form.
+    """
+    form_data: FormData = await request.form()
+    modification_prompt = str(form_data.get("modification_prompt", ""))
+    current_recipe_data = _parse_recipe_form_data(form_data)
+    original_recipe_data = _parse_recipe_form_data(form_data, prefix="original_")
+
+    try:
+        current_recipe = RecipeBase(**current_recipe_data)
+        original_recipe = RecipeBase(**original_recipe_data)
+    except ValidationError as ve:
+        logger.error("Initial validation error in modify recipe: %s", ve)
+        return build_modify_form_response(
+            current_recipe=RecipeBase.model_construct(**current_recipe_data),
+            original_recipe=RecipeBase.model_construct(**original_recipe_data),
+            modification_prompt_value=modification_prompt,
+            error_message_content=Div(
+                "Invalid recipe data. Please check the fields.",
+                cls=f"{CSS_ERROR_CLASS} mt-2",
+            ),
+        )
+
+    if not modification_prompt:
+        logger.debug("Empty modification prompt detected")
+        return build_modify_form_response(
+            current_recipe=current_recipe,
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,  # Preserve entered prompt
+            error_message_content=Div(
+                "Please enter modification instructions.", cls=f"{CSS_ERROR_CLASS} mt-2"
+            ),
+        )
+
+    try:
+        modified_recipe: RecipeBase = await generate_modified_recipe(
+            current_recipe=current_recipe, modification_request=modification_prompt
+        )
+        processed_recipe = postprocess_recipe(modified_recipe)
+        logger.info("LLM modification successful. Building success response.")
+        result = build_modify_form_response(
+            current_recipe=processed_recipe,
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,
+            error_message_content=None,
+        )
+
+    except FileNotFoundError as fnf_e:
+        logger.error("Configuration error - prompt file missing: %s", fnf_e)
+        result = build_modify_form_response(
+            current_recipe=current_recipe,
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,
+            error_message_content=Div(
+                "Service configuration error. Please try again later.",
+                cls=f"{CSS_ERROR_CLASS} mt-2",
+            ),
+        )
+
+    except RuntimeError as llm_e:
+        logger.error("LLM modification error: %s", llm_e)
+        result = build_modify_form_response(
+            current_recipe=current_recipe,  # Revert to recipe before LLM attempt
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,
+            error_message_content=Div(str(llm_e), cls=f"{CSS_ERROR_CLASS} mt-2"),
+        )
+
+    except ValidationError as ve:
+        logger.error("Validation error post-LLM or unexpected: %s", ve)
+        result = build_modify_form_response(
+            current_recipe=current_recipe,  # Revert to recipe before this error
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,
+            error_message_content=Div(
+                "Invalid recipe data after modification attempt.",
+                cls=f"{CSS_ERROR_CLASS} mt-2",
+            ),
+        )
+    except Exception as e:
+        logger.error(
+            "Unexpected error in recipe modification flow: %s", e, exc_info=True
+        )
+        result = build_modify_form_response(
+            current_recipe=current_recipe,  # Revert to recipe before this error
+            original_recipe=original_recipe,
+            modification_prompt_value=modification_prompt,
+            error_message_content=Div(
+                "Critical Error: An unexpected error occurred. "
+                "Please refresh and try again.",
+                cls=CSS_ERROR_CLASS,
+            ),
+        )
     return result
